ETL/requirements/consumer.dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-sources --no-hashes --extra consumer
aiokafka==0.12.0
async-timeout==5.0.1
backoff==2.2.1
certifi==2024.12.14
cffi==1.17.1 ; platform_python_implementation == 'PyPy'
clickhouse-connect==0.8.13
lz4==4.3.3
packaging==24.2
pycparser==2.22 ; platform_python_implementation == 'PyPy'
pytz==2024.2
typing-extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
pydantic-settings==2.9.1
